Lowercase the supported URL schemes, and use scheme list to determine if URL is internal
If a URL has no scheme, it is considered an internal URL. If it has a scheme, we need to lop off the trailing ":" from the protocol and then compare it against the entries in the supported scheme list.

Note that Cocoa does not automatically lowercase the supported schemes (we do this in native though) and Node's url library does.